requirements.txt
pandas==1.3.4
Pillow==8.4.0
plotly==5.5.0
streamlit==1.3.0
yfinance==0.1.67
